Fix bug
